requirements.txt
netcat==0.0.1a0
python-dotenv==1.0.1
textual==0.73.0
